+++ tango/integrations/beaker/executor.py
@@ -461,6 +461,21 @@ class BeakerExecutor(Executor):
         # Ensure entrypoint dataset exists.
         self._ensure_entrypoint_dataset()
 
+        # Get repo info and make sure we're in a GitHub-hosted repository.
+        git = GitMetadata.check_for_repo()
+        if (
+            git is None
+            or git.commit is None
+            or git.remote is None
+            or "github.com" not in git.remote
+        ):
+            raise ExecutorError(
+                f"Missing git data. "
+                f"BeakerExecutor requires a git repository with a GitHub remote."
+            )
+        self._github_account, self._github_repo = self._parse_git_remote(git.remote)
+        self._git_commit = git.commit
+
     def check_repo_state(self):
         if not self.allow_dirty:
             # Make sure repository is clean, if we're in one.
@@ -915,25 +930,11 @@ class BeakerExecutor(Executor):
         experiment_name = (
             f"{Constants.STEP_EXPERIMENT_PREFIX}{step.unique_id}-{str(uuid.uuid4())[:8]}"
         )
-
-        # Ensure we're working in a GitHub repository.
-        git = GitMetadata.check_for_repo()
-        if (
-            git is None
-            or git.commit is None
-            or git.remote is None
-            or "github.com" not in git.remote
-        ):
-            raise ExecutorError(
-                f"Missing git data for step '{step.unique_id}'. "
-                f"BeakerExecutor requires a git repository with a GitHub remote."
-            )
-        try:
-            github_account, github_repo = self._parse_git_remote(git.remote)
-        except ValueError:
-            raise ExecutorError("BeakerExecutor requires a git repository with a GitHub remote.")
-        git_ref = git.commit
-
+        github_account, github_repo, git_ref = (
+            self._github_account,
+            self._github_repo,
+            self._git_commit,
+        )
         if not self._logged_git_info:
             self._logged_git_info = True
             cli_logger.info(