+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+ - Added the `Workspace.remove_step()` method to safely remove steps.
+- The `GSWorkspace()` can now be initialized with google cloud bucket subfolders.
+
+### Changed
+
+- The `BeakerExecutor` now uses the HEAD commit at the time the executor is instantiated to executor a step instead of the HEAD commit at the time the step is run.
+
 ### Fixed
 
 - Removed unnecessary code coverage dev requirements.